4.5 - Vonalhúzás gyakorisága hozzáadva, falnélküli jt_módban nem segfaultol ha átmegyek a pályán

+++ jatek_vezerlo.h
@@ -82,7 +82,7 @@ typedef struct Jatekos {
 typedef struct Vezerles
 {
     bool megallitva_jatek; bool megallitva_felhasznalo;
-    double menetido; bool menet_vege;
+    Uint16 menetido; bool menet_vege;
     int jatekosszam; int max_pontszam; Jatekmod jt_mod;
     double elmozd_jat; double elmozd_lov; double fordulas;
     Falak falak;
